requirements.txt
streamlit==1.24.1
Pillow==9.2.0
easyocr==1.6.2
opencv-python-headless==4.5.5.64
requests==2.31.0
numpy==1.21.6
pandas==1.3.5
